core/executor: don't commit state queue at end of runFunction
Want to be able to inspect initial state and all generated states, but
committing the queue at the end prevents that. Moved it to the
beginning so the result is guaranteed clean from other runs.

notifyCurrent now called commitQueue; alias of stateManager->commitQueue. Had
to move a variable around, etc.